refactor(category page): change in the url parameters and generalization of the API handler
The API handler productQuery now can accept both search query and sort operation requests. It can
also now handle unique product query for when the user clicks on a particular product on the search
page. Improvement in the way all the parameters were being taken from the request url.

BREAKING CHANGE: This affects the way the frontend sends requests to this particular route. It now
has to add a page number arg and a sort by arg. It also affects the product page renders.