feat(base page/category page): addition of sort funtionality(buttons and display)
Introduction of sort buttons and necessary information like the current page number, the number of
products displayed in the current page and the total number of products. Removal of Relevance option
from the sort functionality. Documentation of all the divs and change in url params for the product
page.

BREAKING CHANGE: The frontend sort functionality values have changed which may affect the way the js
files would read and send fetech requests to the backend APIs.

-