fix(category page and backend apis): fix for the changes from backend product-query route
Fix for the way fetch parameter were being sent, introduction of sort functionality through the
backend. Improvement in the way the query params were being extracted from the url. Addition of
pagination functionality which further led to the changes in the url params. Product page can now
fetch unique product data from the backend and render the same on the page. Introduction of a sort
handler to deal with sort functionality and route the page based on the sort order.