perf(docker): addition of the app and all its files onto docker and include redis cache
Following the complete integration b/w the frontend and the backend, as per plan docker images are
created for different services like the flask app, the database and the redis cache. Redis cache has
been included to improve the performance of the entire system. Curretly redis only serves single
product from the database searches. Requirements and config files for the docker to run, have been
created and tested.

BREAKING CHANGE: Major change in both frontend and backend routes and ports